requirements.txt
numpy<2.0
pandas
scikit-learn
scipy
flask
gunicorn
transformers
faiss-cpu
kafka-python
joblib
requests
tqdm